✨ Feat: separate SituationCard component and apply it in situation list
상황 예시 컴포넌트 파일 분리 및 SituationCard 컴포넌트 적용
